Rename bit-os-projects component to os-projects
