--- public/components/bit-os-projects/bit-os-projects.less
@@ -1,3 +1,0 @@
-bit-os-projects {
-	display: block;
-}
