--- public/components/bit-os-projects/bit-os-projects.md
@@ -1,11 +1,0 @@
-@parent bitcentive
-@module {can.Component} bitcentive/components/bit-os-projects <bit-os-projects>
-
-A short description of the bit-os-projects component
-
-@signature `<bit-os-projects>`
-
-@body
-
-## Use
-
