+++ public/components/os-projects/os-projects.html
@@ -1,7 +1,7 @@
 <script type="text/stache" can-autorender id="osProjects">
     <can-import from="bitcentive/models/fixtures/fixtures" />
-    <can-import from="bitcentive/components/bit-os-projects/" />
-    <bit-os-projects {contribution-month}="contributionMonth"/>
+    <can-import from="bitcentive/components/os-projects/" />
+    <os-projects {contribution-month}="contributionMonth"/>
 </script>
 
 <script src="../../../node_modules/steal/steal.js" main="can-view-autorender">
@@ -40,7 +40,7 @@
             clientProject: clientProject
         }]
     });
-    
+
 
     //contributionMonthsListData = contributionMonthStoreFixture.getList();
