+++ public/components/os-projects/os-projects.less
@@ -1,0 +1,3 @@
+os-projects {
+	display: block;
+}
