+++ public/components/os-projects/os-projects.md
@@ -1,0 +1,10 @@
+@parent bitcentive
+@module {can.Component} bitcentive/components/os-projects <os-projects>
+
+A short description of the os-projects component
+
+@signature `<os-projects>`
+
+@body
+
+## Use
