+++ public/test/test.js
@@ -1,4 +1,4 @@
 import 'bitcentive/models/test';
 import 'bitcentive/test/functional';
 
-import 'bitcentive/components/bit-os-projects/bit-os-projects_test';
+import 'bitcentive/components/os-projects/os-projects_test';
